# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This service is to handle My Go ORACLE App. For more detail, please visit https://github.com/",
        "title": "MY GO ORACLE APP API",
        "contact": {
            "name": "Oracle Team",
            "url": "https://github.com/orgs/xxxx/projects/1",
            "email": "oracle.team@mail.com"
        },
        "version": "0.1"
    },
    "basePath": "/my-go-oracle-app",
    "paths": {
        "/members/": {
            "get": {
                "description": "GetAllMembers handles request for Get Member by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Member"
                ],
                "summary": "Get All  by Id",
                "parameters": [
                    {
                        "type": "string",
                        "default": "id",
                        "description": "accept language",
                        "name": "Accept-Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "limit data",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page data",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name filter",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "address filter",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ageStart filter",
                        "name": "ageStart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ageEnd filter",
                        "name": "ageEnd",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "salaryStart filter",
                        "name": "salaryStart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "salaryEnd filter",
                        "name": "salaryEnd",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/oracle_com_oracle_my-go-oracle-app_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/oracle_com_oracle_my-go-oracle-app_service_member.MemberResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "InternalServerError"
                    }
                }
            }
        },
        "/members/{id}": {
            "get": {
                "description": "GetMemberById handles request for Get Member by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Member"
                ],
                "summary": "Get Member by Id",
                "parameters": [
                    {
                        "type": "string",
                        "default": "id",
                        "description": "accept language",
                        "name": "Accept-Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of Member",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/oracle_com_oracle_my-go-oracle-app_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/oracle_com_oracle_my-go-oracle-app_service_member.MemberResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "InternalServerError"
                    }
                }
            }
        }
    },
    "definitions": {
        "oracle_com_oracle_my-go-oracle-app_pkg_response.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "application error code for tracing",
                    "type": "integer",
                    "example": 0
                },
                "msg": {
                    "description": "error message",
                    "type": "string",
                    "example": " "
                },
                "status": {
                    "description": "true if we have error",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "oracle_com_oracle_my-go-oracle-app_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {},
                "data": {},
                "error": {
                    "$ref": "#/definitions/oracle_com_oracle_my-go-oracle-app_pkg_response.Error"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {},
                "serverTime": {
                    "type": "integer"
                }
            }
        },
        "oracle_com_oracle_my-go-oracle-app_service_member.MemberInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "salary": {
                    "type": "integer"
                }
            }
        },
        "oracle_com_oracle_my-go-oracle-app_service_member.MemberResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "info": {
                    "$ref": "#/definitions/oracle_com_oracle_my-go-oracle-app_service_member.MemberInfo"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}